server/econ_model/code/Python/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
DateTime==5.1
econtools==0.3.2
et-xmlfile==1.1.0
fonttools==4.39.4
idna==3.4
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pandas-datareader==0.10.0
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==2.0.2
xlrd==2.0.1
zipp==3.15.0
zope.interface==6.0
